refactor routes for app
