+++ backend-flask/lib/cloudwatch.py
@@ -14,3 +14,7 @@ def init_cloudwatch(response):
     timestamp = strftime('[%Y-%b-%d %H:%M]')
     LOGGER.error('%s %s %s %s %s %s', timestamp, request.remote_addr, request.method, request.scheme, request.full_path, response.status)
     return response
+
+#@app.after_request
+#def after_request(response):
+#  init_cloudwatch(response)
