+++ backend-flask/routes/messages.py
@@ -1,0 +1,56 @@
+import os
+import sys
+
+from flask import current_app as app
+from flask import request, g
+from flask import Flask
+
+from services.message_groups import *
+from services.messages import *
+from services.create_message import *
+
+from lib.cognito_jwt_token import jwt_required
+from lib.helpers import model_json
+from aws_xray_sdk.core import xray_recorder
+from flask_cors import cross_origin
+
+def load(app):
+  @app.route("/api/message_groups", methods=['GET'])
+  @jwt_required()
+  def data_message_groups():
+    model = MessageGroups.run(cognito_user_id=g.cognito_user_id)
+    return model_json(model)
+
+
+  @app.route("/api/messages/<string:message_group_uuid>", methods=['GET'])
+  @jwt_required()
+  def data_messages(message_group_uuid):
+    model = Messages.run(cognito_user_id=g.cognito_user_id,
+    message_group_uuid=message_group_uuid)
+    return model_json(model)
+
+
+  @app.route("/api/messages", methods=['POST','OPTIONS'])
+  @cross_origin()
+  @jwt_required()
+  def data_create_message():
+    message_group_uuid = request.json.get('message_group_uuid',None)
+    user_receiver_handle = request.json.get('handle',None)
+    message = request.json.get('message')
+    if message_group_uuid == None:
+# Create for the first time
+      model = CreateMessage.run(
+      mode="create",
+      message=message,
+      cognito_user_id=g.cognito_user_id,
+      user_receiver_handle=user_receiver_handle
+  )
+    else:
+    # Push onto existing Message Group
+      model = CreateMessage.run(
+      mode="update",
+      message=message,
+      message_group_uuid=message_group_uuid,
+      cognito_user_id=g.cognito_user_id
+      ) 
+    return model_json(model)
